pyproject.toml
[project]
name = "yomutext"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.4",
    "bson>=0.5.10",
    "fastapi[standard]>=0.116.1",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "ruff>=0.12.7",
    "scalar-fastapi>=1.2.3",
    "sqlmodel>=0.0.24",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 120
line.select = ["E", "W", "F", "I"]

